[level 1] Title: 평균 구하기, Time: 0.05 ms, Memory: 86.2 MB -BaekjoonHub

+++ 프로그래머스/1/12944. 평균 구하기/평균 구하기.java
@@ -2,8 +2,8 @@ class Solution {
     public double solution(int[] arr) {
         double answer = 0;
         for (int i = 0; i < arr.length; i++) {
-            answer += arr[i];
+            answer += (double) arr[i];
         }
-        return answer / arr.length;
+        return (double) answer / arr.length;
     }
 }
